% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExampleApps.R
\name{run_06_tabsets}
\alias{run_06_tabsets}
\title{Run the '06 tabsets' example application}
\usage{
run_06_tabsets()
}
\value{
no return value
}
\description{
Run the '06 tabsets' example application
}

